+++ demo/src/App.tsx
@@ -1,16 +1,13 @@
 import * as React from 'react';
 import { Viewer, Worker } from '@react-pdf-viewer/core';
-import { printPlugin } from '@react-pdf-viewer/print';
-import { zoomPlugin } from '@react-pdf-viewer/zoom';
-// import { defaultLayoutPlugin } from '@react-pdf-viewer/default-layout';
+import { defaultLayoutPlugin } from '@react-pdf-viewer/default-layout';
 
 import '@react-pdf-viewer/core/lib/styles/index.css';
 import '@react-pdf-viewer/print/lib/styles/index.css';
-// import '@react-pdf-viewer/default-layout/lib/styles/index.css';
+import '@react-pdf-viewer/default-layout/lib/styles/index.css';
 
 const App = () => {
-    const zoomPluginInstance = zoomPlugin();
-    const printPluginInstance = printPlugin();
+    const defaultLayoutPluginInstance = defaultLayoutPlugin();
 
     return (
         <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.7.570/build/pdf.worker.js">  
@@ -23,8 +20,7 @@ const App = () => {
                 <Viewer
                     fileUrl="http://localhost:8001/pdf-open-parameters.pdf"
                     plugins={[
-                        printPluginInstance,
-                        //zoomPluginInstance,
+                        defaultLayoutPluginInstance,
                     ]}
                 />
             </div>